+++ packages/nx-plugin/src/generators/init/lib/add-analog-dependencies.ts
@@ -9,7 +9,10 @@ export function addAnalogDependencies(
   vitest: boolean,
   nxVersion?: string,
 ) {
-  const devDependencies = getAnalogDevDependencies(angularVersion, nxVersion);
+  const devDependencies: Record<string, string> = getAnalogDevDependencies(
+    angularVersion,
+    nxVersion,
+  );
   const dependencies = getAnalogDependencies(angularVersion);
 
   if (!nxVersion) {