+++ packages/nx-plugin/src/generators/init/lib/update-index-html.ts
@@ -4,13 +4,13 @@ import { SetupAnalogGeneratorSchema } from '../schema';
 
 export function updateIndex(tree: Tree, schema: SetupAnalogGeneratorSchema) {
   const projects = getProjects(tree);
-  const projectConfig = projects.get(schema.project);
+  const projectConfig = projects.get(schema.project)!;
 
   const indexPath = joinPathFragments(projectConfig.root, 'src/index.html');
   const newIndexPath = joinPathFragments(projectConfig.root, 'index.html');
 
   if (tree.exists(indexPath)) {
-    const indexContents = tree.read(indexPath, 'utf-8');
+    const indexContents = tree.read(indexPath, 'utf-8')!;
     let updatedIndex = indexContents.replace(
       '</body>',
       `<script type="module" src="/src/main.ts"></script></body>`,